requirements.txt
colorama==0.4.4
halo==0.0.31
log-symbols==0.0.14
pycryptodome==3.14.1
six==1.16.0
spinners==0.0.24
termcolor==1.1.0
